requirements.txt
Flask
pandas
scikit-learn
Sastrawi
pdfplumber
python-docx
